feat(linux-client): load token from `/etc/dev.firezone.client` ... (#4328)
If it's not in CLI / env var

This is more convenient for development, and it's a step towards getting
the systemd service to work.

The token:
- Can't go in `/usr/lib/systemd/system/firezone-client.service` because
that file is updated by `dpkg`
- Probably shouldn't be in the CLI because CLI args can be seen by other
processes
- Could go in env vars, but those can also be snooped in theory

It has to be stored on disk somewhere for headless operation, so we can
just read it directly from disk.