fix(portal): Fix bug in actor edit page preventing updates (#4347)
Why:

* A bug was present in the actor edit page that prevented updating an
actor due to managed and synced groups being sent as part of the form
submission. Along with that, if a user manually removed the managed
group(s) from the form submission, the actor being edited would be
removed from the managed group, which should not be allowed.

* There was also another small bug which prevent an admin actor from
being updated at all if they were the only admin in the account.