chore(phoenix-channel): reset heartbeat on reconnect (#4361)
Looking through the logs of
https://github.com/firezone/firezone/issues/4348, I noticed that we
would instantly reconnect to the portal due to a "missed heartbeat" if
the connection was reset for any other error. That happens because the
timer within `Heartbeat` was still active and would immediately fire was
soon as we are connected.

To fix this, we introduce a `reset` method that gets called every time
we establish a connection to the portal.