chore: document DNS stub functions (#4526)
Refs #4488 

Part of a yak shave:
- If Windows sees us time out, it will query other DNS servers and
probably cache their response
- If we can return SERVFAIL, I'm not sure if Windows will query other
servers or not
- In order control or even test the stub resolver's behavior, I wanted
to document it first

There's a good chance that even if Windows doesn't cache a SERVFAIL, it
will think that all DNS servers are going to give the same answer and it
will query other servers anyway. Then the problem is not with Windows
caching our response, but with apps caching Windows' response.

Anyway, I have had trouble understanding these functions before, so I
wanted to document them now that I somewhat understand them.

Co-authored-by: Jamil <jamilbk@users.noreply.github.com>