test(ci): Remove e2e false start (#4580)
As we've learned more about how we can test for increased coverage and
certain failure scenarios, I think continuing down this path is a losing
battle.

Apple is the only platform we can't theoretically test in GitHub
actions, and we may be able to accomplish that with #4375. With #4506 in
progress, I think we can get decent coverage with a mix of CI
integration tests and portal-stubbed clients in CI.

If we can stub out the control plane I/O we can test clients in CI.