test(linux-client): separate the token from the systemd unit file
This is needed so that we can auto-update the systemd unit file,
either manually, or with a package manager like `apt`. We don't
want users cut-and-pasting these together on every update, and we don't
want machines doing it. Making the file updatable means we make security
fixes to it easily.