refactor(linux-client): rename `daemon` subcommand to `ipc-service` (#4656)
Closes #4655

This should be more clear since "daemon", like "tunnel", could mean a
variety of things. The IPC thing is the distinct part for this
subcommand, and I didn't want to call it "server" and confuse it with a
web server. "service" hopefully evokes "systemd service" and "Windows
service", something that provides a service locally.

If not it could always be something longer