refactor(connlib): remove `PacketTransform` abstraction (#4709)
As a result of moving all logic into `ClientState` and `GatewayState`,
the concrete types of `Peer` are statically known everywhere. Thus, we
can remove this abstraction layer and directly store a `ClientOnGateway`
and `GatewayOnClient` struct in the `PeerStore`.

This makes code-navigation and reasoning easier because one can directly
jump to the function that is being called.

Resolves: #4224.