Make backoff's per request
This has been bothering me for a while because consecutive requests had different timeouts.
With the introduction of two parallel `BINDING` requests, this actually a problem.

In particular, we currently reset the backoff on every successful response.
This creates a problem if we e.g. don't have an IPv6 interface because the communication over IPv4 will work and will constantly reset the backoff, making the `Allocation` constantly retry the `BINDING` request over IPv6.

To alleviate this, we store a separate backoff per message.