fix(snownet): don't nominate discarded candidates (#4806)
When candidates are being invalidated in str0m, they internally set a
`discarded` flag. This flag was not considered when a candidate that was
previously discarded was added again. This can happen when a client
roams back to their old network, i.e. going from Ethernet to WiFi and
back.

In that case, str0m would still form new pairs using the old discarded
candidate which would also eventually get nominated as part of a pair.
This would lead to a panic in snownet when we try to find the nominated
local candidate as we filter by non-discarded candidates there.

Related: https://github.com/algesten/str0m/pull/508.
Resolves: #4736.

Co-authored-by: Jamil <jamilbk@users.noreply.github.com>